+++ src/main/java/ru/hse/testing/stats/ServerResult.java
@@ -3,15 +3,15 @@ package ru.hse.testing.stats;
 import java.util.List;
 
 public class ServerResult {
-    private final int queryWorkTimeOnServer;
+    private final double queryWorkTimeOnServer;
     private final List<Integer> queryWaitTimeOnClient;
 
-    public ServerResult(int queryWorkTimeOnServer, List<Integer> queryWaitTimeOnClient) {
+    public ServerResult(double queryWorkTimeOnServer, List<Integer> queryWaitTimeOnClient) {
         this.queryWorkTimeOnServer = queryWorkTimeOnServer;
         this.queryWaitTimeOnClient = queryWaitTimeOnClient;
     }
 
-    public int getQueryWorkTimeOnServer() {
+    public double getQueryWorkTimeOnServer() {
         return queryWorkTimeOnServer;
     }
